requirements.txt
openai>=1.12.0
pydantic>=2.5.0
langfuse>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
